requirements.txt
numpy==1.22.4
matplotlib==3.7.1
pandas==2.0.2
rosetta-soil==0.1.1
scikit-learn==1.1.1
seaborn==0.12.2
streamlit==1.15.1
tensorflow==2.9.1
# protobuf==3.19.0
